Initial support for CCClippingNode per #1092
• added new CCClippingNode plugin + icons
• Enhanced drag/drop support for generic nodes
• updated CCBReaderInternal to fix-up UUIDs after deserialization
  (just like Cocos2D CCBReader)
• added unit tests to verify node references are correctly fixed-up
  in Cocos2D CCBReader

Notes:
We probably want a new root node in the timeline view,
for nodes that are not in the main scene graph. It seems like adding
the mask to the scene is not causing any issues, but I'd be interested
in feedback.

As the stencil for a CCClippingNode is not a child,
it's transform is wrong in the GUI. I will look at ways to fix this in
the GUI.